# api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/roles": {
      "post": {
        "operationId": "RolesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      },
      "get": {
        "operationId": "RolesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/roles/{id}": {
      "get": {
        "operationId": "RolesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      },
      "patch": {
        "operationId": "RolesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      },
      "delete": {
        "operationId": "RolesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
...
    "/menus": {
      "post": {
        "operationId": "MenuController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "menus"
        ]
      },
      "get": {
        "operationId": "MenuController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "menus"
        ]
      }
    },
    "/menus/{id}": {
      "get": {
        "operationId": "MenuController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "menus"
        ]
      },
      "patch": {
        "operationId": "MenuController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "menus"
        ]
      },
      "delete": {
        "operationId": "MenuController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "menus"
        ]
      }
    },
    "/menu-routes": {
      "post": {
        "operationId": "MenuRouteController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuRouteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "menu-routes"
        ]
      },
      "get": {
        "operationId": "MenuRouteController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "menu-routes"
        ]
      }
    },
    "/menu-routes/{id}": {
      "get": {
        "operationId": "MenuRouteController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "menu-routes"
        ]
      },
      "patch": {
        "operationId": "MenuRouteController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuRouteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "menu-routes"
        ]
      },
      "delete": {
        "operationId": "MenuRouteController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "menu-routes"
        ]
      }
    },
...
    "/game-type": {
      "post": {
        "operationId": "GameTypesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "game-type"
        ]
      },
      "get": {
        "operationId": "GameTypesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "game-type"
        ]
      }
    },
    "/game-type/{id}": {
      "get": {
        "operationId": "GameTypesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "game-type"
        ]
      },
      "patch": {
        "operationId": "GameTypesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "game-type"
        ]
      },
      "delete": {
        "operationId": "GameTypesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "game-type"
        ]
      }
    },
...
    "/events": {
      "post": {
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "events"
        ]
      },
      "get": {
        "operationId": "EventsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "events"
        ]
      }
    },
    "/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "events"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "events"
        ]
      },
      "delete": {
        "operationId": "EventsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "events"
        ]
      }
    },
...
    "/games/{id}": {
      "get": {
        "operationId": "GamesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "games"
        ]
      },
      "patch": {
        "operationId": "GamesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-admin-bypass",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Admin-Bypass",
            "in": "header",
            "description": "Set to true to bypass \"cannot start if no players\" (admin only)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "games"
        ]
      },
      "delete": {
        "operationId": "GamesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "games"
        ]
      }
    },
...
    "/game-invites/game/{gameId}": {
      "get": {
        "operationId": "GameInvitesController_findByGame",
        "parameters": [
          {
            "name": "gameId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "game-invites"
        ]
      }
    },
...
    "/rewards": {
      "post": {
        "operationId": "RewardsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRewardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "rewards"
        ]
      },
      "get": {
        "operationId": "RewardsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "rewards"
        ]
      }
    },
    "/rewards/{id}": {
      "get": {
        "operationId": "RewardsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "rewards"
        ]
      },
      "patch": {
        "operationId": "RewardsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRewardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "rewards"
        ]
      },
      "delete": {
        "operationId": "RewardsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "rewards"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Unique role name",
            "example": "Admin"
          },
          "description": {
            "type": "string",
            "description": "Optional description of the role",
            "example": "Has full access to all resources"
          },
          "permissions": {
            "description": "List of permission identifiers",
            "example": [
              "user.read",
              "user.write"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Unique role name",
            "example": "Admin"
          },
          "description": {
            "type": "string",
            "description": "Optional description of the role",
            "example": "Has full access to all resources"
          },
          "permissions": {
            "description": "List of permission identifiers",
            "example": [
              "user.read",
              "user.write"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CreateMenuDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name of the menu item",
            "example": "Dashboard"
          },
          "status": {
            "type": "string",
            "description": "Status of the menu item",
            "example": "active"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateMenuDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name of the menu item",
            "example": "Dashboard"
          },
          "status": {
            "type": "string",
            "description": "Status of the menu item",
            "example": "active"
          }
        }
      },
      "CreateMenuRouteDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name of the menu route",
            "example": "Dashboard"
          },
          "url": {
            "type": "string",
            "description": "URL or path this route points to",
            "example": "/dashboard"
          },
          "menu_id": {
            "type": "string",
            "description": "Menu ID this route belongs to",
            "example": "507f1f77bcf86cd799439011"
          },
          "status": {
            "type": "string",
            "description": "Status of the route",
            "example": "active"
          }
        },
        "required": [
          "name",
          "url",
          "menu_id"
        ]
      },
      "UpdateMenuRouteDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name of the menu route",
            "example": "Dashboard"
          },
          "url": {
            "type": "string",
            "description": "URL or path this route points to",
            "example": "/dashboard"
          },
          "menu_id": {
            "type": "string",
            "description": "Menu ID this route belongs to",
            "example": "507f1f77bcf86cd799439011"
          },
          "status": {
            "type": "string",
            "description": "Status of the route",
            "example": "active"
          }
        }
      },
...
      "CreateGameTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the game mode",
            "example": "Arcade"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug for the game mode (auto-generated from name if not provided)",
            "example": "arcade"
          },
          "description": {
            "type": "string",
            "description": "Description of the game mode",
            "example": "Fast-paced arcade-style games"
          },
          "status": {
            "type": "string",
            "description": "Status of the game mode",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateGameTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the game mode",
            "example": "Arcade"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug for the game mode",
            "example": "arcade"
          },
          "description": {
            "type": "string",
            "description": "Description of the game mode",
            "example": "Fast-paced arcade-style games"
          },
          "status": {
            "type": "string",
            "description": "Status of the game mode",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          }
        }
      },
...
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the event",
            "example": "Tech Conference 2024"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug for the event (auto-generated from name if not provided)",
            "example": "tech-conference-2024"
          },
          "category_id": {
            "type": "string",
            "description": "Category ID (references Categories)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "event_type_id": {
            "type": "string",
            "description": "Event Type ID (references EventTypes)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "start_date": {
            "type": "string",
            "description": "Start date/time for the event (ISO string)",
            "example": "2026-03-01T09:00:00Z"
          },
          "end_date": {
            "type": "string",
            "description": "End date/time for the event (ISO string)",
            "example": "2026-03-01T17:00:00Z"
          },
          "eligibility_id": {
            "type": "string",
            "description": "Eligibility ID (references Eligibilities)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "winning_condition": {
            "type": "string",
            "description": "Winning condition (text input)",
            "example": "First to complete 5 rounds"
          },
          "game_type_id": {
            "type": "string",
            "description": "Game Type ID (references Game Types)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "max_participants": {
            "type": "number",
            "description": "Maximum number of participants",
            "example": 100
          },
          "reward_id": {
            "type": "string",
            "description": "Reward ID (references Rewards module)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "rewards_value": {
            "type": "number",
            "description": "Reward value (points, amount, etc.)",
            "example": 500
          },
          "distribution": {
            "type": "string",
            "description": "Rewards distribution method",
            "example": "top-3"
          },
          "status": {
            "type": "string",
            "description": "Status of the event",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "description": {
            "type": "string",
            "description": "Event description",
            "example": "Annual tech conference"
          }
        },
        "required": [
          "name",
          "event_type_id"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the event",
            "example": "Tech Conference 2024"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug for the event",
            "example": "tech-conference-2024"
          },
          "category_id": {
            "type": "string",
            "description": "Category ID (references Categories)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "event_type_id": {
            "type": "string",
            "description": "Event Type ID (references EventTypes)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "start_date": {
            "type": "string",
            "description": "Start date/time for the event (ISO string)",
            "example": "2026-03-01T09:00:00Z"
          },
          "end_date": {
            "type": "string",
            "description": "End date/time for the event (ISO string)",
            "example": "2026-03-01T17:00:00Z"
          },
          "eligibility_id": {
            "type": "string",
            "description": "Eligibility ID (references Eligibilities)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "winning_condition": {
            "type": "string",
            "description": "Winning condition (text input)",
            "example": "First to complete 5 rounds"
          },
          "game_type_id": {
            "type": "string",
            "description": "Game Type ID (references Game Types)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "max_participants": {
            "type": "number",
            "description": "Maximum number of participants",
            "example": 100
          },
          "reward_id": {
            "type": "string",
            "description": "Reward ID (references Rewards module)",
            "example": "665f1a4b2c3d4e5f6a7b8d0e"
          },
          "rewards_value": {
            "type": "number",
            "description": "Reward value (points, amount, etc.)",
            "example": 500
          },
          "distribution": {
            "type": "string",
            "description": "Rewards distribution method",
            "example": "top-3"
          },
          "status": {
            "type": "string",
            "description": "Status of the event",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "description": {
            "type": "string",
            "description": "Event description",
            "example": "Annual tech conference"
          }
        }
      },
...
      "CreateRewardDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name",
            "example": "Gold Trophy"
          },
          "slug": {
            "type": "string",
            "description": "Slug (auto-generated from name if not provided)",
            "example": "gold-trophy"
          },
          "value": {
            "type": "number",
            "description": "Value (e.g. points, amount)",
            "example": 100
          },
          "status": {
            "type": "string",
            "description": "Status",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateRewardDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name",
            "example": "Gold Trophy"
          },
          "slug": {
            "type": "string",
            "description": "Slug",
            "example": "gold-trophy"
          },
          "value": {
            "type": "number",
            "description": "Value",
            "example": 100
          },
          "status": {
            "type": "string",
            "description": "Status",
            "example": "active",
            "enum": [
              "active",
              "inactive"
            ]
          }
        }
      },
...
    }
  }
}